feat: integrate GiftAction component for referral rewards in Perp views
- Added GiftAction component to PerpTickerBarMobile and PerpsHeaderRight for enhanced referral rewards sharing.
- Updated useReferFriends hook to support dynamic source configuration for referral messages.
- Enhanced referral code localization to include Perps-specific titles and subtitles.
