+++ packages/shared/src/referralCode/type.ts
@@ -270,6 +270,18 @@ export interface IInvitePostConfig {
         subtitle: string;
       };
     };
+    Perps?: {
+      title: string;
+      subtitle: string;
+      for_you: {
+        title: string;
+        subtitle: string;
+      };
+      for_your_friend: {
+        title: string;
+        subtitle: string;
+      };
+    };
   };
 }
